split homepage into file per widget
