+++ lib/features/homepage/presentation/widgets/homepage_bullet_text.dart
@@ -1,0 +1,24 @@
+import 'package:flutter/material.dart';
+
+class HomepageBulletText extends StatelessWidget {
+  const HomepageBulletText(this.text, {super.key});
+
+  final String text;
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+    return Row(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        const Text('• '),
+        Expanded(
+          child: Text(
+            text,
+            style: theme.textTheme.bodyMedium,
+          ),
+        ),
+      ],
+    );
+  }
+}
