+++ lib/features/homepage/presentation/widgets/homepage_header.dart
@@ -1,0 +1,28 @@
+import 'package:flutter/material.dart';
+
+class HomepageHeader extends StatelessWidget {
+  const HomepageHeader({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        Text(
+          'Welcome back',
+          style: theme.textTheme.headlineSmall?.copyWith(
+            fontWeight: FontWeight.bold,
+          ),
+        ),
+        const SizedBox(height: 4),
+        Text(
+          'Here is a quick snapshot of your trip',
+          style: theme.textTheme.bodyMedium?.copyWith(
+            color: theme.textTheme.bodyMedium?.color?.withAlpha(180),
+          ),
+        ),
+      ],
+    );
+  }
+}
