+++ lib/features/homepage/presentation/widgets/recommended_for_you_section.dart
@@ -1,0 +1,28 @@
+import 'package:flutter/material.dart';
+import 'package:operation_kalkan/features/homepage/presentation/widgets/homepage_section_card.dart';
+
+class RecommendedForYouSection extends StatelessWidget {
+  const RecommendedForYouSection({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+    return HomepageSectionCard(
+      title: 'Recommended for you',
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          Text(
+            'Check out the new rooftop lounge with sunset tasting menu.',
+            style: theme.textTheme.bodyMedium,
+          ),
+          const SizedBox(height: 8),
+          TextButton(
+            onPressed: () {},
+            child: const Text('View recommendations'),
+          ),
+        ],
+      ),
+    );
+  }
+}
