+++ lib/shared/widgets/safe_network_image.dart
@@ -56,7 +56,7 @@ class SafeNetworkImage extends StatelessWidget {
           return buildPlaceholder();
         },
       );
-    } catch (_) {
+    } on Object {
       return buildPlaceholder();
     }
   }
